requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

python-dotenv

# Machine Learning Utilities
numpy
scikit-learn